Move 4 theme use cases and supporting classes from shared to core-plugin
