+++ androidApp/src/main/kotlin/com/riox432/civitdeck/MainActivity.kt
@@ -18,7 +18,7 @@ import com.riox432.civitdeck.ui.navigation.CivitDeckNavGraph
 import com.riox432.civitdeck.ui.navigation.Tab
 import com.riox432.civitdeck.ui.theme.CivitDeckTheme
 import com.riox432.civitdeck.ui.tutorial.GestureTutorialScreen
-import com.riox432.civitdeck.usecase.GetActiveThemeUseCase
+import com.riox432.civitdeck.plugin.usecase.GetActiveThemeUseCase
 import org.koin.compose.viewmodel.koinViewModel
 import org.koin.core.component.KoinComponent
 import org.koin.core.component.inject
