+++ shared/src/commonMain/kotlin/com/riox432/civitdeck/di/DataModule.kt
@@ -37,10 +37,6 @@ import com.riox432.civitdeck.domain.repository.TagRepository
 import com.riox432.civitdeck.domain.repository.UpdateRepository
 import com.riox432.civitdeck.feature.collections.domain.usecase.ExportWithPluginUseCase
 import com.riox432.civitdeck.feature.collections.domain.usecase.GetAvailableExportFormatsUseCase
-import com.riox432.civitdeck.usecase.ActivateThemePluginUseCase
-import com.riox432.civitdeck.usecase.GetActiveThemeUseCase
-import com.riox432.civitdeck.usecase.ImportThemeUseCase
-import com.riox432.civitdeck.usecase.ObserveThemePluginsUseCase
 import org.koin.core.module.dsl.bind
 import org.koin.core.module.dsl.singleOf
 import org.koin.dsl.module
@@ -74,12 +70,6 @@ val dataModule = module {
     single { KohyaZipExportPlugin(get()) }
     factory { GetAvailableExportFormatsUseCase(get()) }
     factory { ExportWithPluginUseCase(get()) }
-
-    // Theme plugins
-    factory { ImportThemeUseCase(get(), get()) }
-    factory { GetActiveThemeUseCase(get()) }
-    factory { ObserveThemePluginsUseCase(get()) }
-    factory { ActivateThemePluginUseCase(get(), get()) }
 
     // Backup DAO wrappers
     single { CollectionDaos(get()) }
